backend/requirements.txt
# Backend requirements
twikit==2.3.3
postgrest-py
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
python-multipart==0.0.20
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
httpx>=0.25.2
python-dotenv>=0.20.0
beautifulsoup4==4.13.0
pyjwt==2.10.1

# Utility
asyncio>=3.4.3
aiohttp>=3.8.0
tqdm>=4.65.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.19.0
pytest-cov>=3.0.0 